reorganizacion de archivos del componente AdoptarMascota y del formulario de registro
