--- front/src/Componentes/AdoptarPerro/AdoptarPerro.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function AdoptDog () {
-
-    return (
-
-        <div>Listado de perros</div>
-    )
-}
