+++ front/src/Container/App.js
@@ -5,7 +5,7 @@ import HomePage from '../Componentes/HomePage/HomePage';
 import AdoptCat from '../Componentes/AdoptarGato/AdoptarGato';
 import AdoptDog from '../Componentes/AdoptarPerro/AdoptarPerro';
 import InfoProceso from '../Componentes/DonarMascota/InfoProceso';
-import FormRegistro from '../Componentes/DonarMascota/FormRegistro';
+import FormRegistro from '../Componentes/FormRegistro/FormRegistro';
 import Confirmacion from '../Componentes/DonarMascota/Confirmacion';
 import AnimalAbuse from '../Componentes/DenunciarMaltrato/DenunciarMaltrato';
 import OtherThings from '../Componentes/TePuedeInteresar/TePuedeInteresar';
@@ -20,7 +20,7 @@ function App() {
     <Route exact path = '/adoptcat' element = {<AdoptCat/>} />
     <Route exact path = '/adoptdog' element = {<AdoptDog/>} />
     <Route exact path = '/givepet' element = {<InfoProceso/>} />
-    <Route exact path = '/sigin' element = {<FormRegistro/>} />
+    <Route exact path = '/sigup' element = {<FormRegistro/>} />
     <Route exact path = '/confirmation' element = {<Confirmacion/>} />
     <Route exact path = '/animalabuse' element = {<AnimalAbuse/>} />
     <Route exact path = '/otherthings' element = {<OtherThings/>} />
